requirements.txt
asgiref==3.9.1
chardet==5.2.0
click==8.2.1
cmake==4.0.3
colorama==0.4.6
distlib==0.3.9
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib==20.0.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
filelock==3.18.0
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.31.2
wheel==0.45.1
